requirement.txt
numpy>=1.24
tqdm>=4.65
torch>=2.0
torchvision>=0.15
Pillow>=9.0
scikit-image>=0.21
torchxrayvision>=1.2